backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
web3
pydantic
